requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
async-timeout==4.0.3
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.3.0
fakeredis==2.31.0
fastapi==0.116.1
fonttools==4.59.1
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
iniconfig==2.1.0
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain==0.3.27
langchain-community==0.3.28
langchain-core==0.3.75
langchain-openai==0.3.32
langchain-text-splitters==0.3.9
langsmith==0.4.18
marshmallow==3.26.1
matplotlib==3.10.5
multidict==6.6.4
mypy_extensions==1.1.0
numpy==2.2.6
openai==1.101.0
orjson==3.11.3
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pluggy==1.6.0
propcache==0.3.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
regex==2025.7.34
requests==2.32.5
requests-toolbelt==1.0.0
ruff==0.12.10
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.43
starlette==0.47.3
tenacity==9.1.2
tiktoken==0.11.0
tomli==2.2.1
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
yarl==1.20.1
zstandard==0.24.0
